requirements.txt
python-nmap
requests
reportlab
paramiko
asyncssh
asyncio
termcolor
argparse